Move `VdafConfig` to `vdaf` module
