+++ daphne/src/vdaf/prio2.rs
@@ -141,8 +141,8 @@ pub(crate) fn prio2_unshard<M: IntoIterator<Item = Vec<u8>>>(
 #[cfg(test)]
 mod test {
     use crate::{
-        async_test_versions, hpke::HpkeKemId, testing::AggregationJobTest, DapAggregateResult,
-        DapMeasurement, DapVersion, VdafConfig,
+        async_test_versions, hpke::HpkeKemId, testing::AggregationJobTest, vdaf::VdafConfig,
+        DapAggregateResult, DapMeasurement, DapVersion,
     };
 
     async fn roundtrip(version: DapVersion) {
